Compute table of contents.
TOC is represented as a sequence of `TocEntry` values. Each one contains the
id, the title, the section number, and a list of child pages.

The TOC is computed using the data in the page index. To reduce the number of
I/O operations, it reads the metadata to get the page title. Since metadata of
all pages are stored in the same data-block, it should make only a few read
operations.