Integrate `kvlist` into `metadata` module.
Since both `Page` and `Book` are using the same functionality in the `metadata`,
there is no reason to keep a generic implementation of a key-value ser/de
module.

One benefit of this change is that we don't need to create a temporary buffer to
serialize the value of `MetadataEntry::User` entries.

Also, this implementation is much simpler than the previous one, since we don't
need generic support for serializable types (i.e. the `VariantValue` trait).