Use DEFLATE to compress data blocks.
With `BookBuilder::set_compression`, a book can be configured to compress data
blocks with DEFLATE before writing them to a stream.

The decompression is handled transparently, so `DataBlocksReader::with_block`
always returns the original data.

The `writer` example now has a CLI option to create the book with DEFLATE.

This functionality is disabled if the `deflate` feature is not present, which is
available by default.